# docs/swagger.yaml
basePath: /
definitions:
...
  api_handler.ErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/api_handler.FieldError'
        type: array
      message:
        example: Invalid request
        type: string
      status:
        example: 400
        type: integer
    type: object
...
  api_handler.FieldError:
    properties:
      field:
        example: email
        type: string
      message:
        example: Email is required
        type: string
    type: object
  api_handler.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  api_handler.MessageResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
      status:
        example: 200
        type: integer
    type: object
...
  api_handler.PasswordResetSuccessResponse:
    properties:
      message:
        example: If the email exists, a password reset link has been sent
        type: string
      status:
        example: 200
        type: integer
    type: object
...
  api_handler.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
  api_handler.TokenVerificationResponse:
    properties:
      expires_at:
        example: "2025-10-20T12:30:00Z"
        type: string
      message:
        example: Token is valid
        type: string
      valid:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ekklesiahr.com
    name: EkklesiaHR API Support
    url: https://github.com/edsonmubezi/ekklesiahr
  description: Production-ready HR Management System API with comprehensive employee,
    payroll, leave management, and organization administration. Supports multi-tenant
    architecture with role-based access control.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EkklesiaHR API
  version: 1.0.1
paths:
...
  /api/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a password reset email to the user. Returns the same success
        message regardless of whether the email exists (security best practice).
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_handler.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Reset email sent (or email not found)
          schema:
            $ref: '#/definitions/api_handler.PasswordResetSuccessResponse'
        "400":
          description: Invalid email format
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      summary: Request Password Reset
      tags:
      - Authentication
  /api/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Resets user password using a valid reset token. Token can only
        be used once and expires after 30 minutes.
      parameters:
      - description: Reset token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_handler.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
          schema:
            $ref: '#/definitions/api_handler.MessageResponse'
        "400":
          description: Invalid token or weak password
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      summary: Reset Password
      tags:
      - Authentication
  /api/auth/verify-reset-token/{token}:
    get:
      description: Validates a password reset token without consuming it. Use this
        to check token validity before showing password reset form.
      parameters:
      - description: Reset token
        example: '"base64-encoded-token-string"'
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token is valid
          schema:
            $ref: '#/definitions/api_handler.TokenVerificationResponse'
        "400":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      summary: Verify Reset Token
      tags:
      - Authentication
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: 'Provide your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"